requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
opencv-python-headless==4.7.0.72
ultralytics==8.0.120
sqlalchemy==2.0.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
numpy==1.24.3
Pillow==9.5.0